refactor(ui): Refactor organizations/$orgId components
These include all organization statistics cards, plus the products table.
