+++ ui/src/routes/organizations/$orgId/-components/organization-packages-statistics-card.tsx
@@ -17,11 +17,12 @@
  * License-Filename: LICENSE
  */
 
+import { useSuspenseQuery } from '@tanstack/react-query';
 import { Boxes } from 'lucide-react';
 
-import { useOrganizationsServiceGetApiV1OrganizationsByOrganizationIdStatisticsRunsSuspense } from '@/api/queries/suspense';
 import { StatisticsCard } from '@/components/statistics-card';
 import { getEcosystemBackgroundColor } from '@/helpers/get-status-class';
+import { getOrtRunStatisticsByOrganizationIdOptions } from '@/hey-api/@tanstack/react-query.gen';
 import { cn } from '@/lib/utils';
 
 type OrganizationPackagesStatisticsCardProps = {
@@ -33,12 +34,11 @@ export const OrganizationPackagesStatisticsCard = ({
   organizationId,
   className,
 }: OrganizationPackagesStatisticsCardProps) => {
-  const data =
-    useOrganizationsServiceGetApiV1OrganizationsByOrganizationIdStatisticsRunsSuspense(
-      {
-        organizationId: organizationId,
-      }
-    );
+  const data = useSuspenseQuery({
+    ...getOrtRunStatisticsByOrganizationIdOptions({
+      path: { organizationId: organizationId },
+    }),
+  });
 
   const total = data.data.packagesCount;
   const counts = data.data.ecosystems;
